+++ src/main/java/com/developlife/reviewtwits/message/annotation/common/DateFormat.java
@@ -1,7 +1,7 @@
-package com.developlife.reviewtwits.message.annotation.user;
+package com.developlife.reviewtwits.message.annotation.common;
 
 
-import com.developlife.reviewtwits.message.validator.user.BirthdayValidator;
+import com.developlife.reviewtwits.message.validator.user.DateFormatValidator;
 
 import javax.validation.Constraint;
 import java.lang.annotation.ElementType;
@@ -11,8 +11,8 @@ import java.lang.annotation.Target;
 
 @Target(ElementType.FIELD)
 @Retention(RetentionPolicy.RUNTIME)
-@Constraint(validatedBy = BirthdayValidator.class)
-public @interface Birthday {
+@Constraint(validatedBy = DateFormatValidator.class)
+public @interface DateFormat {
     String message() default "생년월일은 현재 날짜보다 이전 날짜여야 하며 yyyy-MM-dd 형식이어야 합니다.";
     Class[] groups() default {};
